requirements.txt
google-generativeai==0.8.3
langchain-google-genai==2.0.4
chromadb==0.5.18
pypdf==5.1.0
langchain==0.3.7
pandas==2.2.3
langchain-community==0.3.5
langchain-chroma==0.1.4